feat: add Redis cache
